requirements.txt
streamlit>=1.25.0
opencv-python-headless
numpy
face_recognition
mediapipe
easyocr
scikit-image
Pillow
torch>=2.0.0
torchvision
transformers